+++ src/hashmap.js
@@ -1,7 +1,7 @@
 /**
  * HashMap - HashMap Implementation for JavaScript
  * @author Jack Moxley <https://github.com/jackmoxley>
- * @version 0.9.0
+ * @version 0.10.0
  * Homepage: https://github.com/mootable/hashmap
  */